+++ HISTORY.md
@@ -9,6 +9,9 @@
 - Check presence of REQUIRED fields
 - Log requests and responses
 - Support defining default mount filesystem
+- Fix UUID warning
+- Show supported CSI spec version on start
+- Support live debugging of running driver
 
 
 ## 0.0.2 (2018-06-19)